test/usb-loopback: write a usb host emulator
Write a mock/emulator for USB host behavior
so that we can test the portable parts of
the USB stack. The emulator runs in a separate
thread from the device, so in principle it can
explore the same data races that a real host
would tickle.

Fix two bugs discovered via testing:

 - A race in usb.c: if a host attempts an IN/OUT
   to quickly after a setup packet, it would get
   a STALL because that was the default state for
   the pipe. Don't set the default state to STALL.
   (And, relatedly, explicitly set it to STALL when
   we can't handle a control transfer.)
 - An issue in usb-cdc-acm.c wherein all DATA packets
   would be 64 bytes.