arch/armv6m: fix getcycles in interrupt context
The armv6m getcycles() needs to handle the case where
the systick counter has underflowed, but the interrupt
has not fired yet because the calling code is executing
in interrupt context.

This bug manifested as the feather-m0/fxs-demo code
hanging after running for a while, because the current
time would jump backwards by one tick, which meant that the
code thought the I2C transaction with the gyroscope never
completed.

Manually tested this fix by streaming a few hundred thousand
records out of the device without any hangs.