madgwick: gyro bias compensation
Unlike the reference Madgwick gyro bias algorithm,
use a low-pass filter of the estimated error to
de-bias gyro readings. Use a very low-bandwidth
filter to ensure we don't bias out 'real' readings,
as the magnetometer readings are pretty noisy.

In my manual testing, this eliminates attitude estimation
drift over periods longer than a second or two. Yaw drift
is hardest to de-bias, probably because the magnetometer
does its own hard-iron compensation. In any case, the
yaw *rate* measurements still appear to be good, and the
yaw measurements are now stable to within a couple degrees
when the system is at rest.